refactor: move apollo setup to plugin
